refactor: extract strength tab to views/StrengthView.jsx
Strangler-fig extraction #2 of the sprint (#75, part of #52). Moves the
~270-line strength tab out of the monolith with no behaviour change.

- new views/StrengthView.jsx owns its local activeLift state, the
  LIFT_COLOR map, and liftColor; strengthTrend + strengthSessions come
  in as props (strengthTrend stays in the monolith — still read by the
  overview's latestSquat)
- monolith renders {view === 'strength' && <StrengthView .../>}; drops
  the now-unused StrengthTooltip / PREHAB_NOTE / STRENGTH_TEMPLATES imports
- render test covers the chart, delta callout, templates, sessions, and
  the single-point empty state; per-file coverage gated at the commercial
  80/80/70 destination (measured 92.8 / 87.5 / 91.7)

Monolith: 6,581 -> 6,303 lines.
